+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.4.0]
+
+- Use webworker so UI remains reponsive
+  - Aborts computation of geometry if file changes during computation instead of queuing
+- Add Content-Security-Policy
+
 ## [0.3.0]
 
 - Add .3mf download